Add files via upload
I am not quite sure if I really should use the exact german translation of "jerkface", which would translate to "Wichser"=a man who is jerking off...
Seems a bit too harsh in my ears. I decided to use "Esel", "jackass" in English. 
Hope that is okay with you guys.