Add files via upload
I had to look up "badger britches" and found out it is some kind of- badger fur trousers? I translated it literally ("Dachsfellhose"), but I'm not sure anybody would know this word. Since it is meant to be an insult, "räudiger Fußabtreter" (="mangy doormat") could be used as an alternative, what do you think?