Relaunch (#152)
* Updated Page 3

* Relaunch changes

Edited according to the new pages, with fixed italics and edited and added dialogue, more info in the changelog.

* Changed <em> to *

* Relaunch changes

Edited according to the new pages, with fixed italics and edited and added dialogue, more info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics and edited and added dialogue, more info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics and edited and added dialogue, more info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. More info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. More info in the changelog. There are <Russian> tags to be filled in later, as well - highlighted in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. More info in the changelog. There are <Russian> tags to be filled in later, as well - highlighted in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. More info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog. There are <Russian> tags to be filled in later, as well - highlighted in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog. There are <Russian> tags to be filled in later, as well - highlighted in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog. There are <Russian> tags to be filled in later, as well - highlighted in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog.

* Relaunch changes

Edited according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog.

* Relaunch changes

Added and edited transcripts according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog.

* Relaunch changes

Added and edited transcripts according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog.

* Relaunch changes (Finished Chapter 4!)

Added and edited transcripts according to the new pages, with fixed italics, bold, as well as edited and added dialogue. Typo fixes. More info in the changelog.

* Ignore .idea/ folder

* Delete erroneous Ukranian.txt

* Moved Francais.txt out of subfolder

* Moved files from 366 to 367

* Moved files from 365 to 366

* Moved remaining files to their new folders

* Fixed some filenames

* Moved misplaced Ukranian.txt

* Fixed some encodings and added ***[OUTDATED]*** to the top of all non-English transcripts up through Page 318

* Type

* Typo

* Removed some endcap typos

---------

Co-authored-by: greeneflower <greeneflower@gmail.com>